Implement a two blocking queues based algorithm to pre-fetch the frames for runs.
In this mechanism, we have two blocking queues one containing the
set of empty prediction buffers and the other containing the set
of read prediction buffers, called full prediction buffers. Whenever
a buffer is available in the empty buffers queue it is taken and
the frame required for the next run that is going to exhaust soon
is read and kept into it. The buffer is pushed to the full buffers
queue. While reading the tuples, the merger fetches the frame as
if they are coming from the file but they are actually the pre-fetched
frames that are kept in full buffer queues. Once such a frame is read
the prediction buffer is taken out of the full queue and pushed
into the empty queue.

The number of predicition buffers available is dictated by the one
who initializes the ExternalSortOperatorDescriptor and hence is
passed as an argument to its constructor.

git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_optimized_sort@1344 123451ca-8445-de46-9d55-352943316053