EC2 revised and bug fix
Only 14.2M of script and jars which be uploaded.

Important bug fix.

If stopnc.sh was not called in the specified directory, then
all files in the filesystem will be deleted. This bug
came from pregelix/pregelix-core/src/main/resources/scripts
and replicated into imru. It should be fixed before we
involved in serious law issue.


git-svn-id: https://hyracks.googlecode.com/svn/branches/fullstack_imru@2870 123451ca-8445-de46-9d55-352943316053